refactor(abw): organize image route helpers
